Файловая структура. Выполнена перегруппировка файлов в проекте Файлы проекта распределены по директориям исходя из их назначения и взаимосвязей
