+++ src/scripts/modules/framework/components/BaseComponent.mjs
