+++ src/scripts/modules/framework/components/Label.mjs
