+++ src/scripts/modules/framework/components/ProgressBar.mjs
