+++ src/scripts/modules/game/assets/background.png
