+++ src/scripts/modules/game/assets/bar.png
