+++ src/scripts/modules/game/assets/bar_back.png
