+++ src/scripts/modules/game/assets/blocks.png
