+++ src/scripts/modules/game/assets/bolt-1-blue.png
