+++ src/scripts/modules/game/assets/bolt-2-purple.png
