+++ src/scripts/modules/game/assets/bolt-3-red.png
