+++ src/scripts/modules/game/assets/bolt-4-yellow.png
