+++ src/scripts/modules/game/assets/bolt-5-green.png
