+++ src/scripts/modules/game/assets/bolt-6-brown.png
