+++ src/scripts/modules/game/assets/button1_base.png
