+++ src/scripts/modules/game/assets/moves.png
