+++ src/scripts/modules/game/assets/pause_base.png
