+++ src/scripts/modules/game/assets/pause_hover.png
