+++ src/scripts/modules/game/assets/pause_press.png
