+++ src/scripts/modules/game/assets/progress_panel.png
