+++ src/scripts/modules/game/assets/score_panel.png
