+++ src/scripts/modules/game/assets/start_background.png
