+++ src/scripts/modules/game/assets/tile.png
