+++ src/scripts/modules/game/assets/top_panel.png
