+++ src/scripts/modules/tools/SpriteSplitter.mjs
